requirements.txt
pyautogui
pygetwindow
psutil
python-dotenv
fastapi
uvicorn[standard]
